+++ Transpilers/Inline/Inline.OpenSCAD.psx.ps1
@@ -46,7 +46,15 @@ if (Shape == "circle") {
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
-$CommandInfo
+$CommandInfo,
+
+# A dictionary of parameters.
+[Collections.IDictionary]
+$Parameter,
+
+# A list of arguments.
+[PSObject[]]
+$ArgumentList
 )
 
 begin {
@@ -60,13 +68,21 @@ begin {
     $startRegex = "(?<PSStart>${IgnoredContext}${startComment}\{$Whitespace)"
     # * EndRegex       ```$whitespace + '}' + $EndComment + $ignoredContext```
     $endRegex   = "(?<PSEnd>$Whitespace\}${endComment}\s{0,}${IgnoredContext})"
+
+    # Create a splat containing arguments to the core inline transpiler
+    $Splat      = [Ordered]@{
+        StartPattern  = $startRegex
+        EndPattern    = $endRegex
+    }
 }
 
 process {
-    
-    $fileInfo = $commandInfo.Source -as [IO.FileInfo]
-    $fileText      = [IO.File]::ReadAllText($fileInfo.Fullname)
+    # Add parameters related to the file
+    $Splat.SourceFile = $commandInfo.Source -as [IO.FileInfo]
+    $Splat.SourceText = [IO.File]::ReadAllText($commandInfo.Source)
+    if ($Parameter) { $splat.Parameter = $Parameter }
+    if ($ArgumentList) { $splat.ArgumentList = $ArgumentList }
 
-    .>PipeScript.Inline -SourceFile $CommandInfo.Source -SourceText $fileText -StartPattern $startRegex -EndPattern $endRegex
+    # Call the core inline transpiler.
+    .>PipeScript.Inline @Splat
 }
-